RDKB-34551:SYS_INFO_WaitingFor_Stack_Init marker is not reported by T2.0
Reason for change:
We report T2 markers only when T2 is true. We will drop the marker if it is in disabled state.
We are using tr181 parameter to read the T2 enable status.
When tr181 is ready to repond your request, It will give true or false based on our T2 configuration.

If you try to read tr181 parameter when it is not even ready, It will give you neither true nor false.
Here t2Shared_api.sh is sourced before tr181 becomes ready. So t2 status is not valid value.
So SYS_INFO_WaitingFor_Stack_Init marker is dropped.

Now we fix this by rechecking the status just before reporting to telemetry
component If the T2 status is invalid (neither true nor false) .

Test Procedure:
Risks: Low
Signed-off-by:Veeraputhiran_Thangavel@comcast.com

Change-Id: Id511b99fb1626f9f59315b905d5593dd29193a77