fix for telemetry GREP markers from any other path than /rdklogs/logs are not getting pushed to ODH ( OFW-888 )
DESCRIPTION:-

The telemetry grep markers from /tmp directory was not getting pushed to ODH since the existing telemetry code will search for the log files in /rdklog/logs path only.

Updated the code to use /rdklogs/logs path as the default path only if the full pat is not specified.

TESTING PERFORMED:-

 1. Enabled telemetry after factory reset.

 2. Verified that the grep markers from /tmp path is getting pushed to ODH with the following content at the DCMResponse
    {
        "header": "EVT_WIFI_CHANNEL_CHANGE_spilt",
        "content": "Channel changed from",
        "type": "/tmp/channel_event_1.log",
        "pollingFrequency": "0"
    },

 3. Verified that the grep markers from /rdklogs/logs are getting uploaded to ODH successfully.