prevent telemetry uploads when device is in Modem mode ( OFW-1990 )
BUG DESCRIPTION:
1. Device is in Dualstack Mode
2. Enabled Telemetry using below commands
     dmcli eRT setv Device.X_LGI-COM_Telemetry.DCMConfigFileURL string https://xxxxx/CH7465LG/DCMresponse-OFW_test.json
     dmcli eRT setv Device.X_LGI-COM_Telemetry.Enable bool true
3. Changed Device mode to Modem Mode from UI
4. Once box is up erouter0 is not having ip address as expected
5. Device is trying to send upload to ODH

Expected Result: When Device is in Modem Mode Telemetry should not send ODH upload.

FIX DESCRIPTION:
If the telemetry is enabled, a condition check added to check whether the device is in bridge mode before telemetry initialise.

TESTING PERFORMED:
Checked the telemetry log. If the device is in bridge mode , the telemetry will exit.

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)