fix Plume not enabled when configured via bootfile ( OFW-2376 F3896SI-713 )
[Issue]
When plume is enabled in the bootfile, opensync does not start after
a factory reset. This is caused by a deadlock during bootup:

A telemetry API called in the WiFi-agent during bootup is doing a get
of the Telemetry.Enable parameter, and PAM is during that time trying
to set the SON URL but it's stuck because it's trying to run
plume_init.sh. plume_init.sh is blocked waiting for the mesh-agent
which still hasn't run, because it's still waiting for the wifi-agent
to initialize

[Solution]
Make the telemetry APIs cache events if the wifi-agent hasn't been
initialized, instead of performing a get of Telemetry.enable. Before
this change, they only waited for PAM to be initialized.