RDKBDEV-1174: replace fgets() with fread() to read the whole DCM config file
Reason for change:
replace fgets() with fread() to read the whole DCM config file
ODH upload is not happening to Correct URL defined in DCM Config File

Regression caused by "RDKB-39831 BCOMB-1083: Fix for httpget hogging issues"

FIX DESCRIPTION:
fgets() to read the DCM configuration saved in HttpOutput.txt stops when it
sees the first newline character. Hence the complete config data is not read
and telemetry goes for the default configuration.
Replaced fgets() with fread() to read the content till the end of file and thus the
whole config data is read without a miss.

Test Procedure:
1. Enable telemetry
2. Ensure that the T2 profile from the DCM configuartion is set successfully
3. Ensure that GPV of Device.X_LGI-COM_Telemetry.UploadRepositoryURL is returning the
   upload URL mentioned in the DCM configuration.

Risks: None.
Signed-off-by: Keerthana V J <kvijayakumarjalajaku.contractor@libertyglobal.com>
Change-Id: I2d4fb0473f9dcc307de870df31e0bab39e1d7268