RDKBDEV-1527: avoid a buffer overflow in initProperties()
Reason for change:
avoid a buffer overflow in initProperties() and updateIncludeConfVal()

The argument to the conversion specifiers "s" should be large
enough to accept the sequence and a terminating null character code,
which shall be added automatically. Here, the buffer "props" size is 255
so the proper length to fscanf() is 254.

Reference:
https://man7.org/linux/man-pages/man3/fscanf.3p.html

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>
Change-Id: Iea425dd5b43011d505875bf745dabab32014d0b7