RDKBDEV-1528: avoid a buffer overflow in getTimezone() and getBuildType()
Reason for change:
Avoid a buffer overflow in getTimezone() and getBuildType()
The argument to the conversion specifiers "s" should be large
enough to accept the sequence and a terminating null character code,
which shall be added automatically. Here, the buffer "fileContent" size is 255
so the proper length to fscanf() is 254.

Reference:
https://man7.org/linux/man-pages/man3/fscanf.3p.html

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>
Change-Id: I0523be03f75439677172ee0f7e0194249cfbf348