fix t2_caching_file keeps increasing ( OFW-4700 F3896SI-2070 )
[BUG DESCRIPTION]]
t2_caching_file keeps increasing on enabling telemetry after factory reset.

[FIX DESCRIPTION]
During factory reset, as there is no local configuration is available in the device, /tmp/.t2ConfigReady is not created.
Because of the absence of this file, telemetry events are getting cached. So added code to create  /tmp/.t2ConfigReady whenever new profile is set.

[TESTING PERFORMED]
1.Factory reset
2.Enable telemetry
  dmcli eRT setv Device.X_LGI-COM_Telemetry.DCMConfigFileURL string http://172.30.117.124/ssl/DCMresponse-OFW_IPv4_NL_test.json
  dmcli eRT setv Device.X_LGI-COM_Telemetry.Enable bool true
3.Make sure that /tmp/t2_caching_file is not increasing.

Signed-off-by: Dhanya Ravindran <dravindran.contractor@libertyglobal.com>