DELIA-64246 RDKB-53273: Observing Increased retry count for Xconf Configuration
Reason for change: To avoid the invalid retry counts for config fetch
and to initialise it zero when it reaches the max value so that when
config fetch fails again it goes for retry correctly.
Test Procedure: Verify that the config fectch takes the old config when
it reached the max limit or above
                For the config failures it counts from the initial value
and goes for max retry  each time and avoid pipe errors.
Risks: Medium
Priority: P1
Signed-off-by: PriyaDharshini_Kathiravan <PriyaDharshini_Kathiravan@Comcast.com>

(cherry picked from commit 2b6d582743998bbf842eaa4c9bac1b8abaa910f6)
Change-Id: Id96106a04e9360be5ae3fbc9e9f3aa88181cdeac