RDKB-53281 RDKB-53295: Add "trim" parameter in multiprofiles parser function
Reason for change: To add the "trim" paramater in multiprofiles struct
so that it fetches the trim value in runtime from the profiles and add
the markervalue.
Test Procedure: Check whether the trim parameter is accepted and value
is added to parameter config.
                Check when the trim param is enabled, leading and
trailing whitespaces should be removed from the markervalue.
Risks: Medium
Priority: P1
Signed-off-by: PriyaDharshini_Kathiravan <PriyaDharshini_Kathiravan@Comcast.com>

RDKB-53616: Adding Gtest test cases for RDKB-53161 US

Reason for change: To add and changes the gtest cases based on the code
changes of RDKB-53161 US
Test Procedure: Check whether the gtest binaries are generated and
working properly
Risks: Low
Priority: P1
Signed-off-by: PriyaDharshini_Kathiravan <PriyaDharshini_Kathiravan@Comcast.com>

Change-Id: I671f624aa40651e35a293244ab8cc63ddbbd1b08