fix wait for wan-status before config file download ( OFW-5498 )
[Issue]:
Telemetry DCM config download is failing when the download is attempted before WAN connectivity is up

[Description]:
When telemetry tries to download DCM config before wan service (erouter0) is up in the Gateway, DCM config download getting failed with the http error "Could not resolve host" and same thing updated in the parameter "Device.X_LGI-COM_Telemetry.DcmDownloadStatus.HTTPStatusString".

[Fix Description]:
Changes added to make the telemetry wait for wan status to become started before trying for DCM config download.

[Testing Performed]:
1. Provision the DUT in DualStack mode.
2. Configure telemetry via bootfile and place the configuration at the beginning of bootfile.
3. Reboot the device.
4. Verify the telemetry download and upload once the device is up.

Signed-off-by: Ravi Kumar K <kravikumar.contractor@libertyglobal.com>

fix zombie processes are observed for CcspPandMSsp component ( OFW-5664 )

[Issue]
Zombie processes are created in PAM after bootup and when performing a
dmcli get of Device.DHCPv6.
This is because PAM is calling _get_shell_output which is defined
differently in both PAM and the telemetry library, and the call is made to the
telemetry _get_shell_output instead.

[Solution]
Change the name of _get_shell_output in telemetry

[Testing performed]
Reboot the gateway
Check that no zombie PAM process is seen
Run dmcli eRT getv Device.DHCPv6.
Check that no zombie PAM process is seen