fix ATOM FREE MEM is decreased to less than 10MB over a period of time ( OFW-5722 )
[Issue]
An increase in RSS memory usage is seen over time for the telemetry2_0 process on both the Atom and ARM sides.
This increase is more significant when large number of WiFi station associations/disassociations occurs.

This is caused by a memory leak in telemetry.

[Solution]
Free the buffer returned by cJSON_PrintUnformatted.

[Testing Performed]
Enable telemetry
Make a station associate/disassociate from the gateway continuously
Verify that the RSS usage for the telemetry2_0 process is not increasing significantly on both the ARM and Atom side