fix Value of upload attempt count is incrementing when http response code is 403 Forbidden error ( OFW-5589 F5685SI-1919 )
[Issue]
value of upload attempt count is showing 0 for every failed attempts and it is not incrementing.

[Description]
Value of upload attempt count is not incrementing when http response code is 403.

[Solution]
modified the logic as to increment the upload attempt count for every failed attempt.

[Testing]
Upgrade the image, and confirm that upload attempt count is incrementing for 403 Forbidden error.

[Upstream Candidate]
Yes

Signed-off-by: Sarveswaran Nagarajan <sanagarajan.contractor@libertyglobal.com>