fix duplicate grep markers are uploaded after loadNewProfile interchip event ( OFW-5889 )
[Issue]
Modem reports duplicate reboot events in telemetry

[Issue Description]
It is observed that the First Seek from the EOF ( firstSeekFromEOF ) is holding non zeo value
in the vector storage of markers in the ATOM telemetry process.
Due to which the last seek pointer of log files are ignored and the entire content of file is considered
while searching for grep marker string. Thus Reboot reason marker ( or any Grep markers ) can get uploaded multiple times.

seekFromEOF option enables the user to specify the amount of the data to
search from the end of the File and should be non zero if and only if we provide it as part of DCM config file.

[Fix Description]
Added missing support to copy firstSeekFromEOF to ATOM telemetry

[Testing Performed]
1. Verify first upload contains Reboot Reason marker
   telemetry.
2. Verify content of /tmp/dca_sorted_file.conf and confirm
   firstSeekFromEOF is 0
3. Verify no further uploads of Reboot Reason marker is seen.

[Upstream Candidate]
Yes, firstSeekFromEOF introduced as part of DELIA-62737, RDKB-50328 :
Enable D_LG_OFW Feature and SeekFromEOF.

Signed-off-by: BijeeshaVS <bvattakaithayilsudha.contractor@libertyglobal.com>