provide blocking_signal variable in http_send app sources ( ** tmp hack ** )
Fix build issues created by upstream change to link libbulkdata with
libxconfclient (which depends on the blocking_signal variable, which
is only defined and set from within the telemetry2_0 app).

Providing an uninitialised variable is a temp hack to fix the build,
it's not expected to work at run time.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>